+++ NumeralSystems.Net/NumeralSystems.Net.Tool/Program.cs
@@ -1,0 +1,7 @@
+namespace NumeralSystems.Net.Tool;
+
+public static class Program
+{
+    public static int Main(string[] args) =>
+        NumsysApplication.Run(args, Console.Out, Console.Error);
+}